+++ NEWS.md
@@ -1,3 +1,4 @@
 # codeInspectr (development version)
 
 * addition of basic functions. Key ones: `plot_function_dependencies()` and `find_reverse_dependencies()`
+* Suite of functions to summarize information about github repositories